Add support for simultaneous heatlayers to map
All heat layers are part of the HeatLayerList. This list also takes care
of blending the opacity values of the layers. The opacity of all but the
last layer is reduced each tick, while the last layers opacity is
increased up to one. Once a layer reaches zero opacity is is removed
from the list.