Change TopicCards to only update when visible
The TopicCards will test if they are visible. If not, the d3 force
simulation will stop. On scroll events the visibility is tested and the
simulation is restarted, when they become visible.

This way, all cards can be shown at once. The Previous and Next Buttons
above and below the list have therefore been removed.