Extend TopicBuffer functionality and adapt classes
The TopicBuffer will now bundle all accesses to a resource in a
RequestList instead of sending multiple JSON get requests. In order
to differentiate requests a key is used that is different for each
type of request. This supports functionalities where each module
grabs the data from the buffer itself instead of getting the data
form its parent.

The Limit buffer has a prune function to reduce the stored elements
until maxLength is reached by searching for the oldest BufferEntry.
Each time a BufferEntry is accessed it will update its lastAccess
value. Therefore, the elements which have not been accessed the
longest will be removed first.