Add mouse input to BarChart and to display
When moving the mouse over a bar chart the bin corresponding to the
mouse position is drawn in a different color. The Chart provides the
callback properties onBinPos and onBinIdx for other classes to react
to user input on the chart. The bin position is the floating point
position of the mouse in a bin, i.e. bin 0 is represented by all
values in [0,1[ and so on. The bin index is just the index of the
selected bin. Either will be null for values outside the current bin
range.