Split wordcloud simulation from TopicCard
The simulation part of the TopicCard has been moved into the WordCloud
component. The topic card handles the loading of data and shows nothing,
a loading symbol or the WordCloud as appropriate. The Wordcloud gets
new data as a prop (topicframe) and creates a TokensSet from this.
The nodes (this.nodes) can be updated. This is done to prepare animation
of the wordcloud when changing the bin/frame for a topic.