Add bin change animation to TopicCard
The bin index is passed from the Page down to the wordcloud as a prop.
When the index is changed the weights of the nodes are updated. When
a token is no longer in the set of tokens, it will get a target
weight of zero. Once the actual weight of the node drops below a
threshold the node will be removed from the list.